Move env_val.rs content to convert.rs
